refactor: Created core package and moved classes into it - Core package will hold items that will most probably not be changed a lot. Core items are also items that will be the base of TSON, they are items that modules would not need to modify and provide a way for modules to interact with other modules - Moved Keyword (and KeywordProvider) to core - Moved ContentProvider to core
