+++ src/main/java/com/euph28/tson/context/TSONContext.java
@@ -10,8 +10,8 @@ import com.euph28.tson.context.provider.StringMapProvider;
 import com.euph28.tson.context.restdata.RequestData;
 import com.euph28.tson.context.restdata.ResponseData;
 import com.euph28.tson.interpreter.TSONInterpreter;
-import com.euph28.tson.interpreter.keyword.Keyword;
-import com.euph28.tson.interpreter.keyword.KeywordProvider;
+import com.euph28.tson.core.keyword.Keyword;
+import com.euph28.tson.core.keyword.KeywordProvider;
 import com.euph28.tson.restclientinterface.TSONRestClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
