+++ src/main/java/com/euph28/tson/core/keyword/Keyword.java
@@ -1,6 +1,7 @@
-package com.euph28.tson.interpreter.keyword;
+package com.euph28.tson.core.keyword;
 
 import com.euph28.tson.context.TSONContext;
+import com.euph28.tson.reporter.TSONReporter;
 
 import java.util.ArrayList;
 import java.util.List;
