+++ src/main/java/com/euph28/tson/core/provider/ContentProvider.java
@@ -1,4 +1,4 @@
-package com.euph28.tson.interpreter.provider;
+package com.euph28.tson.core.provider;
 
 /**
  * Interface for classes that can provide TSON content to the interpreter
